feat: Add support for editable shared links for files
Closes: SDK-2033
