+++ test/integration_new/conftest.py
@@ -1,5 +1,6 @@
+from test.integration_new.context_managers.box_test_user import BoxTestUser
 from test.integration_new.context_managers.local_large_file import LocalLargeFile
-from test.integration_new import util
+from test.integration_new import util, CLIENT
 import pytest
 
 
@@ -42,3 +43,14 @@ def large_file_name():
 def large_file(large_file_name):
     with LocalLargeFile(name=large_file_name) as large_file:
         yield large_file
+
+
+@pytest.fixture(scope="package")
+def other_user():
+    with BoxTestUser(login=None) as other_user:
+        yield other_user
+
+
+@pytest.fixture(scope="package")
+def other_client(other_user):
+    yield CLIENT.as_user(other_user)
