+++ test/unit/object/test_folder.py
@@ -23,6 +23,7 @@ from boxsdk.util.chunked_uploader import ChunkedUploader
 # pylint:disable=protected-access
 # pylint:disable=redefined-outer-name
 from boxsdk.util.datetime_formatter import normalize_date_to_rfc3339_format
+from boxsdk.util.default_arg_value import SDK_VALUE_NOT_SET
 
 
 @pytest.fixture()
@@ -435,6 +436,62 @@ def test_create_subfolder(test_folder, mock_box_session, mock_object_id, mock_fo
     mock_box_session.post.assert_called_once_with(expected_url, data=data)
     assert isinstance(new_folder, Folder)
     assert new_folder.object_id == mock_object_id
+
+
+def test_get_shared_link(
+        test_folder,
+        mock_box_session,
+        shared_link_access,
+        shared_link_unshared_at,
+        shared_link_password,
+        shared_link_can_download,
+        shared_link_can_preview,
+        shared_link_vanity_name,
+        test_url,
+        etag,
+        if_match_header,
+):
+    # pylint:disable=redefined-outer-name, protected-access
+    expected_url = test_folder.get_url()
+    mock_box_session.put.return_value.json.return_value = {
+        'type': test_folder.object_type,
+        'id': test_folder.object_id,
+        'shared_link': {
+            'url': test_url,
+        },
+    }
+    expected_data = {'shared_link': {}}
+    if shared_link_access is not None:
+        expected_data['shared_link']['access'] = shared_link_access
+    if shared_link_unshared_at is not SDK_VALUE_NOT_SET:
+        expected_data['shared_link']['unshared_at'] = normalize_date_to_rfc3339_format(shared_link_unshared_at)
+    if shared_link_can_download is not None or shared_link_can_preview is not None:
+        expected_data['shared_link']['permissions'] = permissions = {}
+        if shared_link_can_download is not None:
+            permissions['can_download'] = shared_link_can_download
+        if shared_link_can_preview is not None:
+            permissions['can_preview'] = shared_link_can_preview
+    if shared_link_password is not None:
+        expected_data['shared_link']['password'] = shared_link_password
+    if shared_link_vanity_name is not None:
+        expected_data['shared_link']['vanity_name'] = shared_link_vanity_name
+
+    url = test_folder.get_shared_link(
+        etag=etag,
+        access=shared_link_access,
+        unshared_at=shared_link_unshared_at,
+        password=shared_link_password,
+        allow_download=shared_link_can_download,
+        allow_preview=shared_link_can_preview,
+        vanity_name=shared_link_vanity_name,
+    )
+    mock_box_session.put.assert_called_once_with(
+        expected_url,
+        data=json.dumps(expected_data),
+        headers=if_match_header,
+        params=None,
+    )
+    assert url == test_url
 
 
 @pytest.mark.parametrize('sync_state', iter(FolderSyncState))
